move interval in separate file
